requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
phonemizer==3.3.0
faiss-cpu==1.13.2
numpy>=1.24,<2.0
python-dotenv==1.1.0
